修复编译警告问题

+++ src/Http/HttpDownloader.h
@@ -48,7 +48,7 @@ public:
 	}
 private:
 	int64_t onResponseHeader(const string &status,const HttpHeader &headers) override;
-	void onResponseBody(const char *buf,size_t size,size_t recvedSize,size_t totalSize) override;
+	void onResponseBody(const char *buf,int64_t size,int64_t recvedSize,int64_t totalSize) override;
 	void onResponseCompleted() override;
 	void onDisconnect(const SockException &ex) override;
     void closeFile();
